Starting a reorg of the code to bring in the diffusion optim to be combined with LML and dry run.
